Logo Component Moved to logos folder
